+++ monitoring/doc.go
@@ -6,11 +6,14 @@
 // by the new BSD license.
 
 // The Tideland Go Library monitoring package supports three kinds of
-// system monitoring.
-//
-// They are helpful to understand what's happening inside a system during
-// runtime. So execution times can be measured and analyzed, stay-set
-// indicators integrated and dynamic control value retrieval provided.
+// system monitoring. They are helpful to understand what's happening
+// inside a system during runtime. So execution times can be measured
+// and analyzed, stay-set variables integrated and dynamic control
+// value retrieval provided. The backend is exchangeable. So the
+// StandardBackend workes like described above, the NullBackend does
+// nothing, and own implementations can integrate external systems.
+// Additionally filters can be added to reduce the monitoring to
+// the points of interest.
 package monitoring
 
 //--------------------
@@ -27,7 +30,7 @@ import (
 
 // PackageVersion returns the version of the version package.
 func PackageVersion() version.Version {
-	return version.New(3, 0, 0)
+	return version.New(4, 0, 0)
 }
 
 // EOF